MINIFICPP-1268 Change the warning level flag to /W3
Also simplify the syntax of these declarations.

Fix byteSwap and unit test it
Use the function wrapper in C++ code; fix the macro in C code.
Fix unreferenced variable warnings
Fix possible loss of data warnings
Fix 'The POSIX name for this item is deprecated' warnings
Set the character encoding to utf-8
Set the source and executable character encoding to utf-8; this gets rid of
many 'character X cannot be represented in the current code page' warnings
in Expression.cpp.

Fix signed/unsigned mismatch warnings

Disable deprecated function warnings
These are deprecated functions delegating to or overriding other
deprecated functions.  They will be removed together, there is no
need for a warning until then.

Replace bitwise | with logical ||
Add missing #includes
Mark the narrowing conversion as narrowing
Change the type of crc_ to uLong


This closes #848

+++ extensions/civetweb/processors/ListenHTTP.cpp
@@ -467,7 +467,7 @@ int64_t ListenHTTP::WriteCallback::process(std::shared_ptr<io::BaseStream> strea
     }
 
     // Transfer buffer data to the output stream
-    stream->write(&buf[0], rlen);
+    stream->write(&buf[0], gsl::narrow<int>(rlen));
 
     nlen += rlen;
   }
