+++ extensions/libarchive/MergeContent.cpp
@@ -98,7 +98,7 @@ std::string MergeContent::readContent(std::string path) {
   std::ifstream in(path.c_str(), std::ios::in | std::ios::binary);
   if (in) {
     in.seekg(0, std::ios::end);
-    contents.resize(in.tellg());
+    contents.resize(gsl::narrow<size_t>(in.tellg()));
     in.seekg(0, std::ios::beg);
     in.read(&contents[0], contents.size());
     in.close();
